usb: host: release the wakelock sources if probe() fails
xHCI driver would register a pair of wakelocks in probe(). But the
pair of wakelocks won't be released until xhci_exynos_remove() is
called. It contributes a error handling problem when probe() fails
without release the wakelocks.

This patch adds the error handle to release wakelocks correctly to
fix this issue.

Bug: 205138535
Signed-off-by: Albert Wang <albertccwang@google.com>
Change-Id: I96e99ae482066fb5f25547dc59fe1f19fb044e94