google/debug: itmon: do not panic on specific speculative accesses
Certain types of access to the ROM space not mapped in the MMU and
couldn’t be traced in the kernel, which is likely speculative access.
The workaround tries to identify such access and avoid counting it
as itmon error.

Bug: 223321349
Change-Id: I6a86ff9a8d9fa3e74c6bdb88a8961565dea4c37a
Signed-off-by: craftju <craftju@google.com>
(cherry picked from commit 6f385a004635cc1597a08f01ed15dc983cc1dafd)