bigocean turn off clocks in idle state
Turn off bigocean clocks when it is idle for predefined timout.
This change is to save the power when user holds onto
bigocean instance but are not actually playing any video.
This can happen with video pause or app bug. The solution is
based on a timer.

Bug: 243066068

Signed-off-by: Vinay Kalia <vinaykalia@google.com>
Signed-off-by: Ruofei Ma <ruofeim@google.com>
Change-Id: Ib7a09de489b8443821f1ef6b2f57d32dd5d9740a
(cherry picked from commit 979c1bdc6048890c841e8771840fceaea9238267)